pyproject.toml
[tool.poetry]
name = "repl-nix-mustyyearlyapplicationprogrammer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_mustyyearlyapplicationprogrammer"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
aiohttp = "3.11.10"
discord-py = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
